+++ leadgen-app-form.php
@@ -3,7 +3,7 @@
  * Plugin Name: LeadGen App Form Plugin
  * Plugin URI: https://github.com/SilverAssist/leadgen-app-form
  * Description: WordPress plugin that adds a shortcode to display LeadGen App forms with desktop-id and mobile-id parameters.
- * Version: 1.3.0
+ * Version: 1.3.1
  * Author: Silver Assist
  * Author URI: http://silverassist.com/
  * Text Domain: leadgen-app-form
@@ -13,7 +13,7 @@
  * License URI: https://polyformproject.org/licenses/noncommercial/1.0.0/
  *
  * @package LeadGenAppForm
- * @version 1.3.0
+ * @version 1.3.1
  * @author Silver Assist
  */
 
@@ -25,7 +25,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 // Define plugin constants.
-define( 'LEADGEN_APP_FORM_VERSION', '1.3.0' );
+define( 'LEADGEN_APP_FORM_VERSION', '1.3.1' );
 define( 'LEADGEN_APP_FORM_FILE', __FILE__ );
 define( 'LEADGEN_APP_FORM_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 define( 'LEADGEN_APP_FORM_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );